refactor: split detached-glass module into feature submodules
Mirror the glass/ refactor: break the monolithic detached-glass/module.js
into per-feature submodules (add, activate, move, resize, restore) and fold
the aggregator directly into index.js, so module.js is no longer needed.

Move/resize interaction state is now closure-local instead of leaking onto
the BinaryWindow instance via the spread. bringToFront and its topZIndex
counter move onto DetachedGlassManager, co-located with the active-glass
registry they manage. Drop an unused manager import from minimize.js.
