+++ src/binary-window/detached-glass/index.js
@@ -1,7 +1,26 @@
+import { default as addModule } from './add';
+import { default as activateModule } from './activate';
+import { default as moveModule } from './move';
+import { default as resizeModule } from './resize';
+import { default as restoreModule } from './restore';
 import { closeAction } from './close';
 import { minimizeAction } from './minimize';
 
 export { DetachedGlass } from './class';
-export { default } from './module';
 
 export const BUILTIN_ACTIONS_2 = [minimizeAction, closeAction];
+
+export default {
+  enableDetachedGlassFeatures() {
+    this.enableDetachedGlassActivate();
+    this.enableDetachedGlassResize();
+    this.enableDetachedGlassMove();
+    this.handleMinimizedDetachedGlassClick();
+  },
+
+  ...addModule,
+  ...activateModule,
+  ...moveModule,
+  ...resizeModule,
+  ...restoreModule,
+};
